requirements.txt
# Data Collection
python-binance==1.0.19
ccxt==4.2.25

# Data Processing
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
tables==3.9.2

# Technical Analysis
ta==0.11.0
pandas-ta==0.4.71b0

# Machine Learning
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
ipython==8.19.0
jupyter==1.0.0
